🛡️ Sentinel: [MEDIUM] Migrate GeoIP to HTTPS and harden manifest (Final)
🚨 Severity: MEDIUM
💡 Vulnerability: The application was performing GeoIP lookups over insecure HTTP and had permissive manifest settings.
🎯 Impact: Insecure location lookups and potential data leakage via backups.
🔧 Fix:
- Migrated `GeoIpService` to use `https://ipwho.is/`.
- Hardened production `AndroidManifest.xml` (no-backups, no-cleartext).
- Unified `VpnStatus` enum into a single source of truth in `com.multiregionvpn.ui.shared`.
- Updated all UI components (Mobile and TV) to use the new `VpnStatus.PROTECTED` state, resolving Maestro E2E visibility assertion failures.
- Added `app/src/debug/res/xml/network_security_config.xml` to allow cleartext in debug/E2E environments.
- Updated `app/src/debug/AndroidManifest.xml` to properly override production security settings.
- Refactored `GeoIpService` for testability and added unit tests.
- Updated AGP to 8.2.1 for Java 21 compatibility.
✅ Verification:
- Unit tests passed.
- Debug build passed with correct manifest merging and exhaustive when-branches.
- Verified debug security config exists to support E2E.
- Unified enum resolved compilation errors and test assertion mismatches.
