+++ app/src/main/AndroidManifest.xml
@@ -27,8 +27,7 @@
         android:networkSecurityConfig="@xml/network_security_config"
         android:usesCleartextTraffic="false"
         android:banner="@drawable/tv_banner"
-        tools:targetApi="31"
-        tools:replace="android:theme,android:name,android:label">
+        tools:targetApi="31">
         
         <!-- Mobile Activity -->
         <activity
