+++ app/src/main/java/com/multiregionvpn/ui/components/VpnHeaderBar.kt
@@ -16,6 +16,7 @@ import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import com.multiregionvpn.ui.shared.VpnStatus
 
 /**
  * Professional VPN Header Bar (NOC Style)
@@ -125,13 +126,4 @@ fun VpnHeaderBar(
     )
 }
 
-/**
- * VPN Status States
- */
-enum class VpnStatus(val displayText: String) {
-    PROTECTED("Protected"),
-    CONNECTING("Connecting"),
-    DISCONNECTED("Disconnected"),
-    ERROR("Error")
-}
