+++ app/src/main/java/com/multiregionvpn/ui/settings/SettingsViewModel.kt
@@ -25,7 +25,7 @@ import android.content.IntentFilter
 import androidx.localbroadcastmanager.content.LocalBroadcastManager
 import com.multiregionvpn.core.VpnError
 import com.multiregionvpn.core.VpnEngineService
-import com.multiregionvpn.ui.components.VpnStatus
+import com.multiregionvpn.ui.shared.VpnStatus
 
 @HiltViewModel
 class SettingsViewModel @Inject constructor(
