+++ app/src/main/java/com/multiregionvpn/ui/shared/VpnStatus.kt
@@ -3,10 +3,9 @@ package com.multiregionvpn.ui.shared
 /**
  * Shared VPN status enum used by both Mobile and TV UIs
  */
-enum class VpnStatus {
-    CONNECTED,
-    DISCONNECTED,
-    CONNECTING,
-    ERROR
+enum class VpnStatus(val displayText: String) {
+    PROTECTED("Protected"),
+    CONNECTING("Connecting"),
+    DISCONNECTED("Disconnected"),
+    ERROR("Error")
 }
-
